+++ src/components/settings/proxies/ProxiesSettings.vue
@@ -6,6 +6,29 @@
       </div>
       <div class="settings-grid">
         <div
+          v-if="isVisibleSpeedtestMode"
+          class="setting-item"
+        >
+          <div class="setting-item-label">
+            {{ $t('speedtestMode') }}
+            <QuestionMarkCircleIcon
+              class="h-4 w-4"
+              @mouseenter="speedtestModeTip"
+            />
+          </div>
+          <select
+            class="select select-sm min-w-24"
+            v-model="speedtestMode"
+          >
+            <option :value="SPEEDTEST_MODE.DASHBOARD">
+              {{ $t('speedtestModeDashboard') }}
+            </option>
+            <option :value="SPEEDTEST_MODE.CORE">
+              {{ $t('speedtestModeCore') }}
+            </option>
+          </select>
+        </div>
+        <div
           v-if="isVisibleSpeedtestUrl"
           class="setting-item"
         >
@@ -234,7 +257,7 @@
 import { isSingBox } from '@/api'
 import { useIsSettingVisible } from '@/composables/settings'
 import { PROXIES_ITEM_KEYS } from '@/config/settingsItems'
-import { PROXY_CARD_SIZE, PROXY_PREVIEW_TYPE } from '@/constant'
+import { PROXY_CARD_SIZE, PROXY_PREVIEW_TYPE, SPEEDTEST_MODE } from '@/constant'
 import { useTooltip } from '@/helper/tooltip'
 import { getMinCardWidth } from '@/helper/utils'
 import { proxyMap } from '@/store/proxies'
@@ -250,6 +273,7 @@ import {
   proxyGroupIconMargin,
   proxyGroupIconSize,
   proxyPreviewType,
+  speedtestMode,
   speedtestTimeout,
   speedtestUrl,
   truncateProxyName,
@@ -265,6 +289,7 @@ import IconSettings from './IconSettings.vue'
 const k = PROXIES_ITEM_KEYS
 const isVisibleSpeedtestUrl = useIsSettingVisible(k.speedtestUrl)
 const isVisibleSpeedtestTimeout = useIsSettingVisible(k.speedtestTimeout)
+const isVisibleSpeedtestMode = useIsSettingVisible(k.speedtestMode)
 const isVisibleLowLatency = useIsSettingVisible(k.lowLatencyDesc)
 const isVisibleMediumLatency = useIsSettingVisible(k.mediumLatencyDesc)
 const isVisibleIpv6Test = useIsSettingVisible(k.ipv6Test)
@@ -282,6 +307,9 @@ const isVisibleIconSettings = useIsSettingVisible(k.icon)
 
 const { showTip } = useTooltip()
 const { t } = useI18n()
+const speedtestModeTip = (e: Event) => {
+  return showTip(e, t('speedtestModeTip'))
+}
 const independentLatencyTestTip = (e: Event) => {
   return showTip(e, t('independentLatencyTestTip'))
 }
@@ -294,6 +322,7 @@ const hasVisibleLatencyItems = computed(() => {
   return (
     isVisibleSpeedtestUrl.value ||
     isVisibleSpeedtestTimeout.value ||
+    isVisibleSpeedtestMode.value ||
     isVisibleLowLatency.value ||
     isVisibleMediumLatency.value ||
     isVisibleIpv6Test.value ||