+++ src/constant/index.ts
@@ -74,6 +74,16 @@ export enum PROXY_PREVIEW_TYPE {
   BAR = 'bar',
 }
 
+export enum PROXY_SEARCH_MODE {
+  GLOBAL = 'global',
+  GROUP = 'group',
+}
+
+export enum SPEEDTEST_MODE {
+  CORE = 'core',
+  DASHBOARD = 'dashboard',
+}
+
 export enum CONNECTION_DISPLAY_STYLE {
   AUTO = 'auto',
   CARD = 'card',