+++ src/store/settings.ts
@@ -14,8 +14,10 @@ import {
   PROXY_CARD_SIZE,
   PROXY_CHAIN_DIRECTION,
   PROXY_PREVIEW_TYPE,
+  PROXY_SEARCH_MODE,
   PROXY_SORT_TYPE,
   SETTINGS_MENU_KEY,
+  SPEEDTEST_MODE,
   TABLE_SIZE,
   TABLE_WIDTH_MODE,
   TEST_URL,
@@ -219,6 +221,18 @@ export const twoColumnProxyGroup = useStorage('config/two-columns', true)
 export const speedtestUrl = useStorage<string>('config/speedtest-url', TEST_URL)
 export const independentLatencyTest = useStorage('config/independent-latency-test', false)
 export const speedtestTimeout = useStorage<number>('config/speedtest-timeout', 5000)
+export const speedtestMode = useStorage<SPEEDTEST_MODE>(
+  'config/speedtest-mode',
+  SPEEDTEST_MODE.DASHBOARD,
+)
+export const proxySearchMode = useStorage<PROXY_SEARCH_MODE>(
+  'config/proxy-search-mode',
+  PROXY_SEARCH_MODE.GROUP,
+)
+export const proxyProviderSearchMode = useStorage<PROXY_SEARCH_MODE>(
+  'config/proxy-provider-search-mode',
+  PROXY_SEARCH_MODE.GLOBAL,
+)
 export const proxySortType = useStorage<PROXY_SORT_TYPE>(
   'config/proxy-sort-type',
   PROXY_SORT_TYPE.DEFAULT,